+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Version 1.55
 
-- Fix for Angular projects that are missing .browserslistrc
+- When releases are prepared it will now open the folder the ipa/apk/aab file that was generated and may open store portal
+- Fix for Angular projects that are missing .browserslistrc (fixes webpack error on some Angular projects in iOS 15)
 
 ### Version 1.54