+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - When releases are prepared it will now open the folder the ipa/apk/aab file that was generated and may open store portal
 - NX support for project.json with comments
+- Support for folder based monorepo's that use a different package manager than npm
 - Fix for Angular projects that are missing .browserslistrc (fixes webpack error on some Angular projects in iOS 15)
 
 ### Version 1.54